backend/requirements.txt
Django==5.1.8
strawberry-graphql==0.263.1
strawberry-graphql-django==0.57.1
strawberry-django-auth==0.378.4
psycopg[binary]==3.2.6
celery==5.5.1
redis==5.2.1
django-celery-beat==2.7.0
